requirements-web.txt
# StreamVault Pro Web Dependencies
flet>=0.27.6
flet-video==0.1.0
httpx>=0.28.1
screeninfo>=0.8.1
aiofiles>=24.1.0
streamget>=4.0.5
python-dotenv>=1.0.1
cachetools>=5.5.2
psutil>=6.0.0
Pillow>=10.4.0
pydantic>=2.8.2
pydantic-settings>=2.4.0
requests>=2.32.3
websockets>=12.0
aliyun-python-sdk-core>=2.13.36
tencentcloud-sdk-python-asr==3.0.1394
oss2>=2.18.4
cos-python-sdk-v5>=1.9.30
cryptography>=41.0.0